+++ docs/source/example-code/src/test/java/net/corda/docs/java/tutorial/testdsl/CommercialPaperTest.java
@@ -4,11 +4,9 @@ import kotlin.Unit;
 import net.corda.core.contracts.PartyAndReference;
 import net.corda.core.contracts.TransactionVerificationException;
 import net.corda.core.identity.CordaX500Name;
-import net.corda.core.internal.GlobalProperties;
 import net.corda.finance.contracts.ICommercialPaperState;
 import net.corda.finance.contracts.JavaCommercialPaper;
 import net.corda.finance.contracts.asset.Cash;
-import net.corda.testing.common.internal.ParametersUtilities;
 import net.corda.testing.core.TestIdentity;
 import net.corda.testing.node.MockServices;
 import org.junit.Before;
@@ -17,7 +15,6 @@ import org.junit.Test;
 import java.security.PublicKey;
 import java.time.temporal.ChronoUnit;
 
-import static java.util.Collections.emptyList;
 import static java.util.Collections.singletonList;
 import static net.corda.core.crypto.Crypto.generateKeyPair;
 import static net.corda.finance.Currencies.DOLLARS;
@@ -36,8 +33,7 @@ public class CommercialPaperTest {
     private MockServices ledgerServices;
 
     @Before
-    public void setUp() throws Exception {
-        GlobalProperties.setNetworkParameters(ParametersUtilities.testNetworkParameters(emptyList()));
+    public void setUp() {
         // When creating the MockServices, you need to specify the packages that will contain the contracts you will use in this test
         // For this test its' Cash and CommercialPaper, bot in the 'net.corda.finance.contracts' package
         // In case you don't specify the 'cordappPackages' argument, the MockServices will be initialised with the Contracts found in the package of the current test