+++ node/src/integration-test/kotlin/net/corda/node/services/BFTNotaryServiceTests.kt
@@ -27,7 +27,7 @@ import net.corda.nodeapi.internal.DevIdentityGenerator
 import net.corda.nodeapi.internal.network.NetworkParametersCopier
 import net.corda.core.node.NotaryInfo
 import net.corda.testing.core.chooseIdentity
-import net.corda.testing.common.internal.ParametersUtilities.testNetworkParameters
+import net.corda.testing.common.internal.testNetworkParameters
 import net.corda.testing.contracts.DummyContract
 import net.corda.testing.core.dummyCommand
 import net.corda.testing.node.MockNodeParameters